% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betterKUenm.R
\docType{package}
\name{betterKUenm}
\alias{betterKUenm}
\title{betterKUenm: a better version of kuenm}
\description{
This is the description for the package. It is very informative.
}
\details{
Lots and lots of details here.
}
\section{Main functions in biosurvey}{

\code{\link{ext_back}}, \code{\link{fitGLM}},
\code{\link{occ_prep}}, \code{\link{projectionGLM}}
}

\section{Other functions (important helpers)}{

}

\section{Data included}{

}

