% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_back.R
\name{ext_back}
\alias{ext_back}
\title{Extract background data from variables}
\usage{
ext_back(variables, M = NULL, n)
}
\arguments{
\item{variables}{(SpatRaster) SpatRaster with layers to be used in the model}

\item{M}{(Optional SpatVector) #Polygon delimiting the calibration area}

\item{n}{(numeric) #Number of pixels to extract}
}
\value{
A dataframe containing the values extracted from variables
}
\description{
ext_back mask the environmental variables and extract the values of n cells at random
}
