% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGLM.R
\name{fitGLM}
\alias{fitGLM}
\title{Fitting Generalized Linear Models for only presences data}
\usage{
fitGLM(presences, background, w_presences = 1, w_backg = 1000)
}
\arguments{
\item{presences}{matrix of environmental values for presences records}

\item{background}{matrix of environmental values for background records}

\item{w_presences}{a numeric value to weight presence records}

\item{w_backg}{a numeric value to weight background records}
}
\value{
fitGLM returns an object of class inheriting from "glm" which inherits from the class "lm".
}
\description{
Fitting Generalized Linear Models for only presences data
}
