% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{glm_model}
\alias{glm_model}
\title{Example of a glm model calibrated with records and variables}
\format{
glm object:
}
\usage{
glm_model
}
\description{
Example of a glm model calibrated with records and variables
}
\examples{
data("glm_model", package = "betterKUenm")

summary(glm_model)
}
\keyword{datasets}
