% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\name{M}
\alias{M}
\title{Example of accessible areas for a species}
\format{
A SpatVector object.
}
\value{
No return value, used with function \code{\link[terra]{vect}} to
bring an example of an accessible area for a species.
}
\description{
A SpatVector object representing the accessible area for a species.
}
\examples{
M <- terra::vect(system.file("extdata", "M.gpkg", package = "betterKUenm"))

terra::plot(M)
}
