% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_prep.R
\name{occ_prep}
\alias{occ_prep}
\title{Title Extract variables in coordinates}
\usage{
occ_prep(data, lon, lat, variables)
}
\arguments{
\item{data}{a matrix with at least 2 columns containing longitude and latitude data}

\item{lon}{longitude}

\item{lat}{latitude}

\item{variables}{SpatRaster with values to be extracted for the coordinates}
}
\value{
matrix
}
\description{
Title Extract variables in coordinates
}
