% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectionGLM.R
\name{projectionGLM}
\alias{projectionGLM}
\title{Project GLMs to Raster Variables}
\usage{
projectionGLM(model, new_data)
}
\arguments{
\item{model}{GLM results with weights}

\item{new_data}{A SpatRaster file}
}
\value{
Projected raster
}
\description{
Project GLMs to Raster Variables
}
\examples{
model <- glm(formula=pa~., data = v)
new_data <- terra::rast("./ex/raster.tif")
projectionGLM(model, new_data)
}
