% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{records}
\alias{records}
\title{Example of occurrence records for one species}
\format{
data.frame:
\describe{
\item{species}{species name, a code in this example}
\item{longitude}{longitude, longitude value}
\item{latitude}{latitude, latitude value}
}
}
\usage{
records
}
\description{
Example of occurrence records for one species
}
\examples{
data("records", package = "betterKUenm")

str(records)
}
\keyword{datasets}
