% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\name{variables}
\alias{variables}
\title{Example of environmental variables used in analysis}
\format{
A SpatRaster object.
}
\value{
No return value, used with function \code{\link[terra]{rast}} to
bring an example of an environmental variable used in analysis.
}
\description{
A SpatRaster object representing the variable temperature.
}
\examples{
variables <- terra::rast(system.file("extdata", "variables.tif", package = "betterKUenm"))

terra::plot(variables)
}
